Update to support Subprotocols. Solves #639 (#642)
Ocelot websocket middleware did not work for STOMP over websocket. After investigation i found out that the issue was with subprotocol and headers that are send and filtered. 

I the end i used ASP.Net core proxy as a reference to solve the issue here:

https://github.com/aspnet/Proxy/blob/3015029f5169b699dfa16bcf85cfc7ca0aaddae7/src/Microsoft.AspNetCore.Proxy/ProxyAdvancedExtensions.cs

So i modified the code to use the way ASP.Net proxy handles this.